fix a datatype on test_htable and make msvc build more strict
darcs-hash:20061121141523-16063-49869c94adf22bc9aae54477a316f199f7dd212e.gz

+++ libsofia-sip-ua/su/test_htable.c
@@ -158,7 +158,7 @@ void zap(context_t *c, entry_t *e)
 static int count(context_t *c, hash_value_t h)
 {
   entry_t *e, **ee;
-  int n;
+  unsigned long n;
 
   for (ee = htable_hash(c->c_hash, h), n = 0;
        (e = *ee);
